Move each test helper into its own file inside a new __test-helpers folder.
The intent is:

1. Make it obvious what test-helpers we have by making them visible in the file system
2. Make it clear where they live (__tests__ is for test files
