+++ packages/shopify-app-remix/src/server/__test-helpers/__tests__/request-mock.test.ts
@@ -3,7 +3,7 @@ import {
   mockExternalRequests,
   skipMockChecks,
   validateMocks,
-} from './request-mock';
+} from '../request-mock';
 
 describe('Request mocks', () => {
   beforeEach(() => {
