+++ packages/shopify-app-remix/src/server/authenticate/admin/__tests__/respond-to-options.test.ts
@@ -1,9 +1,5 @@
 import {shopifyApp} from '../../..';
-import {
-  APP_URL,
-  getThrownResponse,
-  testConfig,
-} from '../../../__tests__/test-helper';
+import {APP_URL, getThrownResponse, testConfig} from '../../../__test-helpers';
 import {REAUTH_URL_HEADER} from '../../const';
 
 describe('authorize.admin', () => {
