+++ packages/shopify-app-remix/src/server/authenticate/public/appProxy/__tests__/authenticate.test.ts
@@ -8,8 +8,8 @@ import {
   getThrownResponse,
   setUpValidSession,
   testConfig,
-} from '../../../../__tests__/test-helper';
-import {mockExternalRequest} from '../../../../__tests__/request-mock';
+  mockExternalRequest,
+} from '../../../../__test-helpers';
 
 const REQUEST_URL = `https://${TEST_SHOP}/admin/api/${LATEST_API_VERSION}/customers.json`;
